Rename clients
Making this refactor for incoming rekorclient code
